VoiceBot/requirements.txt
streamlit==1.28.1
groq==0.4.1
python-dotenv==1.0.0
speechrecognition==3.10.0
pyttsx3==2.90
pyaudio==0.2.14
streamlit-mic-recorder==0.0.8
audio-recorder-streamlit==0.0.8
streamlit-webrtc==0.47.1
numpy==1.26.4
soundfile==0.12.1